Refactor `lib_ROI_overlap.py` (ref #616)
