+++ fractal_tasks_core/lib_regions_of_interest/__init__.py
@@ -1,2 +1,3 @@
 from .v1 import *  # noqa
 from .v1_checks import *  # noqa
+from .v1_overlaps import *  # noqa
