+++ fractal_tasks_core/tasks/create_ome_zarr.py
@@ -30,7 +30,7 @@ from fractal_tasks_core.lib_metadata_parsing import parse_yokogawa_metadata
 from fractal_tasks_core.lib_parse_filename_metadata import parse_filename
 from fractal_tasks_core.lib_regions_of_interest import prepare_FOV_ROI_table
 from fractal_tasks_core.lib_regions_of_interest import prepare_well_ROI_table
-from fractal_tasks_core.lib_ROI_overlaps import remove_FOV_overlaps
+from fractal_tasks_core.lib_regions_of_interest import remove_FOV_overlaps
 from fractal_tasks_core.lib_tables import write_table
 from fractal_tasks_core.lib_write import open_zarr_group_with_overwrite
