+++ tests/test_unit_ROIs.py
@@ -26,6 +26,9 @@ from fractal_tasks_core.lib_regions_of_interest import (
     convert_ROIs_from_3D_to_2D,
 )
 from fractal_tasks_core.lib_regions_of_interest import empty_bounding_box_table
+from fractal_tasks_core.lib_regions_of_interest import (
+    find_overlaps_in_ROI_indices,
+)
 from fractal_tasks_core.lib_regions_of_interest import get_image_grid_ROIs
 from fractal_tasks_core.lib_regions_of_interest import get_single_image_ROI
 from fractal_tasks_core.lib_regions_of_interest import is_ROI_table_valid
@@ -36,7 +39,6 @@ from fractal_tasks_core.lib_regions_of_interest import load_region
 from fractal_tasks_core.lib_regions_of_interest import prepare_FOV_ROI_table
 from fractal_tasks_core.lib_regions_of_interest import prepare_well_ROI_table
 from fractal_tasks_core.lib_regions_of_interest import reset_origin
-from fractal_tasks_core.lib_ROI_overlaps import find_overlaps_in_ROI_indices
 
 
 PIXEL_SIZE_X = 0.1625
