+++ tests/test_unit_parse_yokogawa_metadata.py
@@ -20,8 +20,8 @@ from devtools import debug
 from pandas import Timestamp
 
 from fractal_tasks_core.lib_metadata_parsing import parse_yokogawa_metadata
-from fractal_tasks_core.lib_ROI_overlaps import remove_FOV_overlaps
-from fractal_tasks_core.lib_ROI_overlaps import run_overlap_check
+from fractal_tasks_core.lib_regions_of_interest import remove_FOV_overlaps
+from fractal_tasks_core.lib_regions_of_interest import run_overlap_check
 
 # General variables and paths (relative to the test folder)
 testdir = os.path.dirname(__file__)
